+++ src/fkooman/X509/CertParser.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2013 François Kooman <fkooman@tuxed.net>
+ * Copyright 2015 François Kooman <fkooman@tuxed.net>.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,8 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 namespace fkooman\X509;
+
+use InvalidArgumentException;
+use RuntimeException;
 
 class CertParser
 {
@@ -31,7 +33,7 @@ class CertParser
     public function __construct($certData)
     {
         if (!is_string($certData)) {
-            throw new CertParserException("input should be string");
+            throw new InvalidArgumentException('input should be string');
         }
 
         $pattern = '/-----BEGIN CERTIFICATE-----(.*)-----END CERTIFICATE-----/msU';
@@ -40,20 +42,20 @@ class CertParser
         }
 
         // create one long string of the certificate
-        $replaceCharacters = array(" ", "\t", "\n", "\r", "\0" , "\x0B");
+        $replaceCharacters = array(' ', "\t", "\n", "\r", "\0" , "\x0B");
         $certData = str_replace($replaceCharacters, '', $certData);
 
         // store this stripped certificate
         $this->strippedCert = $certData;
 
         // parse the certificate using OpenSSL
-        if (!function_exists("openssl_x509_parse")) {
-            throw new CertParserException("php openssl extension not available");
+        if (!function_exists('openssl_x509_parse')) {
+            throw new RuntimeException('OpenSSL extension not available');
         }
 
         $c = openssl_x509_parse($this->toPEM());
         if (false === $c) {
-            throw new CertParserException("unable to parse the certificate");
+            throw new InvalidArgumentException('unable to parse the certificate');
         }
 
         $this->parsedCert = $c;
@@ -74,14 +76,14 @@ class CertParser
         // prepend header and append footer
         $wrapped = wordwrap($this->toBase64(), 64, "\n", true);
 
-        return "-----BEGIN CERTIFICATE-----".PHP_EOL.$wrapped.PHP_EOL."-----END CERTIFICATE-----".PHP_EOL;
+        return '-----BEGIN CERTIFICATE-----'.PHP_EOL.$wrapped.PHP_EOL.'-----END CERTIFICATE-----'.PHP_EOL;
     }
 
     public static function fromFile($fileName)
     {
         $fileData = @file_get_contents($fileName);
         if (false === $fileData) {
-            throw new CertParserException("unable to read file");
+            throw new RuntimeException('unable to read file');
         }
 
         return new static($fileData);
@@ -92,8 +94,8 @@ class CertParser
      */
     public function getNotValidBefore()
     {
-        if (!array_key_exists("validFrom_time_t", $this->parsedCert)) {
-            throw new CertParserException("could not find 'validFrom_time_t' key");
+        if (!array_key_exists('validFrom_time_t', $this->parsedCert)) {
+            throw new RuntimeException('could not find "validFrom_time_t" key');
         }
 
         return $this->parsedCert['validFrom_time_t'];
@@ -104,17 +106,17 @@ class CertParser
      */
     public function getNotValidAfter()
     {
-        if (!array_key_exists("validTo_time_t", $this->parsedCert)) {
-            throw new CertParserException("could not find 'validTo_time_t' key");
+        if (!array_key_exists('validTo_time_t', $this->parsedCert)) {
+            throw new RuntimeException('could not find "validTo_time_t" key');
         }
 
         return $this->parsedCert['validTo_time_t'];
     }
 
-    public function getFingerprint($algorithm = "sha1", $uriSafe = false)
+    public function getFingerprint($algorithm = 'sha256', $uriSafe = false)
     {
         if (!in_array($algorithm, hash_algos())) {
-            throw new CertParserException(sprintf("unsupported algorithm '%s'", $algorithm));
+            throw new RuntimeException(sprintf('unsupported algorithm "%s"', $algorithm));
         }
 
         if ($uriSafe) {
@@ -134,55 +136,58 @@ class CertParser
     }
 
     /**
-     * Get the common name
+     * Get the common name.
      */
     public function getName()
     {
-        if (!array_key_exists("name", $this->parsedCert)) {
-            throw new CertParserException("could not find 'name' key");
+        if (!array_key_exists('name', $this->parsedCert)) {
+            throw new RuntimeException('could not find "name" key');
         }
 
         return $this->parsedCert['name'];
     }
 
     /**
-     * Get the whole subject as string
+     * Get the whole subject as string.
      *
      * @throws CertParserException
+     *
      * @return string
      */
     public function getSubject()
     {
         // @codeCoverageIgnoreStart
         if (!array_key_exists('subject', $this->parsedCert)) {
-            throw new CertParserException("could not find 'subject' key");
+            throw new RuntimeException('could not find "subject" key');
         }
         // @codeCoverageIgnoreEnd
         return $this->toDistinguishedName($this->parsedCert['subject']);
     }
 
     /**
-     * Get the whole subject as string
+     * Get the whole subject as string.
      *
      * @throws CertParserException
+     *
      * @return string
      */
     public function getIssuer()
     {
         // @codeCoverageIgnoreStart
         if (!array_key_exists('issuer', $this->parsedCert)) {
-            throw new CertParserException("could not find 'issuer' key");
+            throw new RuntimeException('could not find "issuer" key');
         }
         // @codeCoverageIgnoreEnd
         return $this->toDistinguishedName($this->parsedCert['issuer']);
     }
 
     /**
-     * Checks whether current cert is issued by given cert
+     * Checks whether current cert is issued by given cert.
      *
      * @param CertParser $cert
      *
      * @throws CertParserException
+     *
      * @return bool
      */
     public function isIssuedBy(CertParser $cert)
@@ -191,7 +196,7 @@ class CertParser
     }
 
     /**
-     * Returns parsed array data
+     * Returns parsed array data.
      *
      * @return array
      */
@@ -201,7 +206,7 @@ class CertParser
     }
 
     /**
-     * Transforms the array notification of the distinguished name component to string
+     * Transforms the array notification of the distinguished name component to string.
      *
      * @param array  $data
      * @param string $separator
@@ -214,10 +219,10 @@ class CertParser
         foreach ($data as $key => $item) {
             if (is_array($item)) {
                 foreach ($item as $value) {
-                    $output [] = "$key=$value";
+                    $output[] = sprintf('%s=%s', $key, $value);
                 }
             } else {
-                $output [] = "$key=$item";
+                $output[] = sprintf('%s=%s', $key, $item);
             }
         }